Split include/otf2/event/metric.hpp
Move the typed_value_proxy, value_proxy and metric_values classes into
their own separate files in include/otf2xx/event/detail.
