+++ src/reader/callback/events.cpp
@@ -143,7 +143,7 @@ namespace reader
                     otf2::chrono::convert(reader->ticks_per_second())(otf2::chrono::ticks(
                         time - reader->clock_properties().start_time().count()));
 
-                otf2::event::metric::metric_values metric_values{
+                otf2::event::metric::values metric_values{
                     std::vector<OTF2_Type>{ typeIDs, typeIDs + numberOfMetrics },
                     std::vector<OTF2_MetricValue>{ metricValues, metricValues + numberOfMetrics }
                 };
